moving learning and txt folders
